feat(firmware): 新增 get_controls 串口命令
- serial_config_protocol.h: 新 op get_controls，可选 reset 布尔字段
- serial_config.ino: sendControlCounts 上报 34 控件计数 + midi_rx + 当前音符，reset 时清零计数（帧长最坏 ~400 字节，512 上限内）
- Wingie2.ino: 控件计数改为 uint16（单次产测不可能超过 65535 次，保证响应帧不溢出）
- tests/host: 覆盖 get_controls 有无 reset 解析与非法 reset 拒绝

编译通过，host 测试与 93 项 pytest 全绿。

+++ Wingie2/Wingie2.ino
@@ -150,11 +150,11 @@ int potValRealtime[3], potValSampled[3], midi_ch_l, midi_ch_r, midi_ch_both, use
 volatile uint32_t midi_rx_count = 0; // 成功解析的 MIDI 消息总数，用于串口 status 观测 TRS 接收是否存活
 
 struct ControlActivity {
-  volatile uint32_t key[2][12];      // 24 键
-  volatile uint32_t modeButton[2];   // 2 模式键
-  volatile uint32_t octButton[2][2]; // 4 八度键
-  volatile uint32_t sourceSwitch;    // 音源开关
-  volatile uint32_t pot[3];          // 3 电位器
+  volatile uint16_t key[2][12];      // 24 键
+  volatile uint16_t modeButton[2];   // 2 模式键
+  volatile uint16_t octButton[2][2]; // 4 八度键
+  volatile uint16_t sourceSwitch;    // 音源开关
+  volatile uint16_t pot[3];          // 3 电位器
 };
 
 volatile ControlActivity controlActivity;
